More PISWAP Fixes.
PISWAP should be functional in 80-90% of cases.  Some networks (like yeast-human) completely break, for some reason.  I suspect this is because of proteins with the same name in both networks, but I can't be sure.  I'll investigate this if it poses a bigger problem than just yeast-human.

Also, removed PISWAP's autocorrection ability so that it will actually accept whatever alpha you give it instead of deciding it knows better than you.

Main fix is adding a static function in Alignment that will read an edgelist as an alignment with no respect to order (nodes in G1 can be in either column, as long as the nodes in G2 are in the other for that particular row).  This might be useful in the future, so I thought I'd make it clear what I added.


Former-commit-id: bc69bfac934a133b9b9b5ef4470654fdeae56067