refactored temperature schedule and added new methods (#88)
* typos in helpOutput and minor improvements

* refactored automatic temp schedule selector

Also added place holders for the two new automatic methods. Also fixed a bug where TDecay is not set when -tintial is explicit and -tdecay is by-linear-regression.

* temperature schedule methods refactored

renamed methods more uniformly and accurately, removed (a lot of) dead code

* cleaned up getPBad function

cleaning the function in preparation for the more sophisticated 'getPBadAtEquilibrium' version

* reimplemented getpbad based on equilibrium

it seems that the recent history of accepted pbad values is not informative enough to determine if we are at equilibrium. Next commit, I will try keeping track of the score. that should work better

* getpbad detects thermal eq. based on the score trend

* unified doubling methods

several methods based on linear search in log space have been put together in one consistent method. also: changed the behavior of the binary search within L.R. because it had a gap where if 'mid' fell in it it would not get updated anymore and keep getting pbad for the same temp

* pBad Binary Search for temp schedule

added method to set initial temp by binary search using getPBad

* preliminary implementation of Ameur method

not tested yet, just a transcription from the paper

* finished ameur method

it might still require fine tuning

* improved termination condition for Ameur method

probably still requires additional finetuning