hybrid bin-search+LR method, superLR for method comparison, bug fixes
now the method comparison can report on the results of binary search and LR based on the same runs of binary search. It also does LR on all the samples taken by all the methods (superLR). This commit also fixes a bug in Ameur method (errorTolerance and startGuess were passed in wrong order).